Whole Project completed
Registration and log-in for users are being done and authentication is
being done by the  Spring security,All the CRUD operations for
Product,Supplier,Category and every thing is created and updated in
database and the cart,shipping is created. The images are being loaded
to the project location and their corresponding values are being
updated in the database, that means they are kind of being loaded from
database in the page , user can shop the certain product and add it to
the cart and proceed to check out. Where as Admin can do all the works
like Adding , deleting products, suppliers,categories etc.,,